+++ frontend/src/django-generator/ProModal.js
@@ -27,9 +27,8 @@ export function ProModal({ open, handleClose, price }) {
                                     </div>
                                     <p className="text-gray-500 dark:text-gray-300 text-sm md:text-base">
                                         <span className="text-gray-900 dark:text-white">
-                                            <a target="_blank" href="/onboarding-kit/">Onboarding Kit</a>
-                                        </span>
-                                        - Premium Dashboards, Dynamic Django Tool, and eCommerce CMS included - <a target="_blank" href="/onboarding-kit/">See Details</a>
+                                            <a target="_blank" href="/onboarding-kit/">Onboarding Kit</a>{' - '}
+                                        </span>Premium Dashboards, Dynamic Django Tool, and eCommerce CMS included - <a target="_blank" href="/onboarding-kit/">See Details</a>.
                                     </p>
                                 </div>
                             </div>
@@ -44,7 +43,7 @@ export function ProModal({ open, handleClose, price }) {
                                     <p className="text-gray-500 dark:text-gray-300 text-sm md:text-base">
                                         <span className="text-gray-900 dark:text-white font-bold">
                                             Unlock a New Premium Starter at each 4mo
-                                        </span> (Personal License)
+                                        </span> (Personal License) - <a target="_blank" href="/product/?search=pro">See List</a>.
                                     </p>
                                 </div>
                             </div>
@@ -59,8 +58,7 @@ export function ProModal({ open, handleClose, price }) {
                                     <p className="text-gray-500 dark:text-gray-300 text-sm md:text-base">
                                         <span className="text-gray-900 dark:text-white font-bold">
                                             Premium Support
-                                        </span>
-                                        - Monday-Sunday, 3h response time.
+                                        </span>{' - '}Monday-Sunday, 3h response time.
                                     </p>
                                 </div>
                             </div>
@@ -74,9 +72,8 @@ export function ProModal({ open, handleClose, price }) {
                                     </div>
                                     <p className="text-gray-500 dark:text-gray-300 text-sm md:text-base">
                                         <span className="text-gray-900 dark:text-white">
-                                            <a href="/tools/django-generator/">Django Generator</a>
-                                        </span>
-                                        Unlimited starters, all features available.
+                                            <a target="_blank" href="/tools/django-generator/">Django Generator</a>{' - '}
+                                        </span>Unlimited starters, all features available.
                                     </p>
                                 </div>
                             </div>
@@ -90,9 +87,23 @@ export function ProModal({ open, handleClose, price }) {
                                     </div>
                                     <p className="text-gray-500 dark:text-gray-300 text-sm md:text-base">
                                         <span className="text-gray-900 dark:text-white">
-                                            <a href="/tools/csv-processor/">CSV Processor</a>
-                                        </span>
-                                        Full access (all features).
+                                            <a target="_blank" href="/tools/flask-generator/">Flask Generator</a>{' - '}
+                                        </span>Unlimited starters, all features available.
+                                    </p>
+                                </div>
+                            </div>                            
+
+                            <div className="flex flex-col gap-3 md:gap-3 mb-3">
+                                <div className="flex items-start gap-2">
+                                    <div className="bg-primary-100 rounded-full p-0.5">
+                                        <svg className="w-5 h-5 text-primary-700" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" width="24" height="24" fill="none" viewBox="0 0 24 24">
+                                            <path stroke="currentColor" strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M5 11.917 9.724 16.5 19 7.5" />
+                                        </svg>
+                                    </div>
+                                    <p className="text-gray-500 dark:text-gray-300 text-sm md:text-base">
+                                        <span className="text-gray-900 dark:text-white">
+                                            <a target="_blank" href="/tools/csv-processor/">CSV Processor</a>{' - '}
+                                        </span>Full Access (all features).
                                     </p>
                                 </div>
                             </div>
@@ -106,9 +117,8 @@ export function ProModal({ open, handleClose, price }) {
                                     </div>
                                     <p className="text-gray-500 dark:text-gray-300 text-sm md:text-base">
                                         <span className="text-gray-900 dark:text-white">
-                                            <a href="/tools/db-migrator/">DataBase Migrator</a>
-                                        </span>
-                                        Full access (all features).
+                                            <a target="_blank" href="/tools/db-migrator/">DataBase Migrator</a>{' - '}
+                                        </span>Full access (all features).
                                     </p>
                                 </div>
                             </div>
@@ -122,9 +132,8 @@ export function ProModal({ open, handleClose, price }) {
                                     </div>
                                     <p className="text-gray-500 dark:text-gray-300 text-sm md:text-base">
                                         <span className="text-gray-900 dark:text-white">
-                                            <a href="/tools/db-processor/">DataBase Processor</a>
-                                        </span>
-                                        Full access (all features).
+                                            <a target="_blank" href="/tools/db-processor/">DataBase Processor</a>{' - '}
+                                        </span>Full Access (all features).
                                     </p>
                                 </div>
                             </div>
@@ -138,7 +147,7 @@ export function ProModal({ open, handleClose, price }) {
                                 </h2>
                                 <p className="text-gray-500 dark:text-gray-300 text-xs md:text-sm">Monthly subscription (cancel anytime)</p>
                             </div>
-                            <a href="https://appseed.gumroad.com/l/pro-subscription" className="mb-5 text-white w-full inline-flex items-center justify-center bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm px-5 py-2.5 text-center dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800">
+                            <a target="_blank" href="https://appseed.gumroad.com/l/pro-subscription" className="mb-5 text-white w-full inline-flex items-center justify-center bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm px-5 py-2.5 text-center dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800">
                                 Subscribe
                                 <svg className="w-5 h-5 text-white ml-2" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" width="24" height="24" fill="none" viewBox="0 0 24 24">
                                     <path stroke="currentColor" strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M19 12H5m14 0-4 4m4-4-4-4" />
@@ -147,7 +156,7 @@ export function ProModal({ open, handleClose, price }) {
                             <div className="text-center w-full">
                                 <p className="text-gray-500 dark:text-gray-300 text-xs md:text-sm">
                                     Payment Secured by</p>
-                                <img className="block mx-auto" src="" alt="" />
+                                <img className="block mx-auto" src="/static/dist/img/gumroad.png" alt="GUMROAD Image" />
                             </div>
                         </div>
                     </div>